move tests to root folder; change rena related imports in tests utils to local imports
