+++ tests/ReplayTest.py
@@ -20,9 +20,9 @@ AppConfigs(_reset=True)  # create the singleton app configs object
 
 from rena.presets.Presets import PresetType
 from rena.config import stream_availability_wait_time
-from rena.tests.TestStream import LSLTestStream
-from rena.tests.test_utils import get_random_test_stream_names, app_fixture, ContextBot
-from rena.utils.data_utils import RNStream
+from tests.TestStream import LSLTestStream
+from tests.test_utils import get_random_test_stream_names, app_fixture, ContextBot
+from rena.utils.RNStream import RNStream
 from rena.utils.ui_utils import CustomDialog
