+++ tests/StartupTest.py
@@ -9,7 +9,7 @@ import pytest
 from rena.configs.configs import AppConfigs
 AppConfigs(_reset=True)  # create the singleton app configs object
 
-from rena.tests.test_utils import app_fixture, ContextBot
+from tests.test_utils import app_fixture, ContextBot
 
 
 @pytest.fixture
