+++ tests/TestStream.py
