+++ tests/archived/UnityCameraCaptureTest.py
