+++ tests/test_utils.py
@@ -16,16 +16,8 @@ from PyQt6.QtCore import Qt
 from PyQt6.QtWidgets import QWidget
 from pytestqt.qtbot import QtBot
 
-from rena.MainWindow import MainWindow
-from rena.config import stream_availability_wait_time
-from rena.presets.Presets import PresetType
-from rena.startup import load_settings, apply_patches
-from rena.sub_process.pyzmq_utils import can_connect_to_port
-from rena.tests.TestStream import LSLTestStream, ZMQTestStream
-from rena.tests.test_viz import visualize_metrics_across_num_chan_sampling_rate
-from rena.utils.buffers import flatten
-from rena.utils.data_utils import RNStream
-from rena.utils.ui_utils import CustomDialog
+from tests.TestStream import LSLTestStream, ZMQTestStream
+from tests.test_viz import visualize_metrics_across_num_chan_sampling_rate
 
 
 def app_fixture(qtbot, show_window=True, revert_to_default=True, reload_presets=True):
@@ -34,8 +26,11 @@ def app_fixture(qtbot, show_window=True, revert_to_default=True, reload_presets=
     print(os.getcwd())
     # ignore the splash screen and tree icon
     app = QtWidgets.QApplication(sys.argv)
+    from rena.startup import load_settings
     load_settings(revert_to_default=revert_to_default, reload_presets=reload_presets)  # load the default settings
+    from rena.startup import apply_patches
     apply_patches()
+    from rena.MainWindow import MainWindow
     test_renalabapp_main_window = MainWindow(app=app, ask_to_close=False)  # close without asking so we don't pend on human input at the end of each function test fixatire
     if show_window:
         test_renalabapp_main_window.show()
@@ -43,11 +38,11 @@ def app_fixture(qtbot, show_window=True, revert_to_default=True, reload_presets=
 
     return app, test_renalabapp_main_window
 
-def stream_is_available(app: MainWindow, test_stream_name: str):
+def stream_is_available(app, test_stream_name: str):
     # print(f"Stream name {test_stream_name} availability is {app.stream_widgets[test_stream_name].is_stream_available}")
     assert app.stream_widgets[test_stream_name].is_stream_available
 
-def streams_are_available(app: MainWindow, test_stream_names: List[str]):
+def streams_are_available(app, test_stream_names: List[str]):
     # print(f"Stream name {test_stream_name} availability is {app.stream_widgets[test_stream_name].is_stream_available}")
     for ts_name in test_stream_names:
         assert app.stream_widgets[ts_name].is_stream_available
@@ -56,6 +51,7 @@ def stream_is_unavailable(app_main_window, stream_name):
     assert not app_main_window.stream_widgets[stream_name].is_stream_available
 
 def handle_custom_dialog_ok(qtbot, patience_second=0, click_delay_second=0):
+    from rena.utils.ui_utils import CustomDialog
     if patience_second == 0:
         w = QtWidgets.QApplication.activeWindow()
         if isinstance(w, CustomDialog):
@@ -72,7 +68,9 @@ def handle_custom_dialog_ok(qtbot, patience_second=0, click_delay_second=0):
         print(f": {w} is a CustomDialog, trying to click ok button")
         yes_button = w.buttonBox.button(QtWidgets.QDialogButtonBox.StandardButton.Ok)
         qtbot.mouseClick(yes_button, QtCore.Qt.MouseButton.LeftButton, delay=int(click_delay_second * 1e3))
-def handle_current_dialog_ok(app: MainWindow, qtbot: QtBot, patience_second=0, click_delay_second=0):
+
+
+def handle_current_dialog_ok(app, qtbot: QtBot, patience_second=0, click_delay_second=0):
     """
     This is compatible with CustomDialogue creation that also sets the current_dialog in MainWindow
     @param app: the main window
@@ -82,7 +80,7 @@ def handle_current_dialog_ok(app: MainWindow, qtbot: QtBot, patience_second=0, c
     """
     handle_current_dialog_button(QtWidgets.QDialogButtonBox.StandardButton.Ok, app, qtbot, patience_second, click_delay_second)
 
-def handle_current_dialog_button(button, app: MainWindow, qtbot: QtBot, patience_second=0, click_delay_second=0):
+def handle_current_dialog_button(button, app, qtbot: QtBot, patience_second=0, click_delay_second=0):
     """
     This is compatible with CustomDialogue creation that also sets the current_dialog in MainWindow
     @param app: the main window
@@ -90,6 +88,8 @@ def handle_current_dialog_button(button, app: MainWindow, qtbot: QtBot, patience
     @param patience_second: how long to wait for the current dialog to be a CustomDialog
     @param delay: how long to wait before clicking the button
     """
+    from rena.utils.ui_utils import CustomDialog
+
     if patience_second == 0:
         if isinstance(app.current_dialog, CustomDialog):
             yes_button = app.current_dialog.buttonBox.button(button)
@@ -113,11 +113,12 @@ class ContextBot:
     Helper class for carrying out the most performed actions in the tests
 
     """
-    def __init__(self, app: MainWindow, qtbot: QtBot):
+    def __init__(self, app, qtbot: QtBot):
         self.send_data_processes = {}
         self.app = app
         self.qtbot = qtbot
 
+        from rena.config import stream_availability_wait_time
         self.stream_availability_timeout = int(20 * stream_availability_wait_time * 1e3)
 
         self.monitor_stream_name = "monitor 0"
@@ -137,6 +138,7 @@ class ContextBot:
         p = Process(target=LSLTestStream, args=(stream_name, num_channels, srate))
         p.start()
         self.send_data_processes[stream_name] = p
+        from rena.presets.Presets import PresetType
         self.app.create_preset(stream_name, PresetType.LSL, num_channels=num_channels, nominal_sample_rate=srate)  # add a default preset
 
         self.app.ui.tabWidget.setCurrentWidget(self.app.ui.tabWidget.findChild(QWidget, 'visualization_tab'))  # switch to the visualization widget
@@ -149,6 +151,7 @@ class ContextBot:
         self.qtbot.mouseClick(self.app.stream_widgets[stream_name].StartStopStreamBtn, QtCore.Qt.MouseButton.LeftButton)
 
     def create_zmq_stream(self, stream_name: str, num_channels: int, srate:int, port_range=(5000, 5100)):
+        from rena.sub_process.pyzmq_utils import can_connect_to_port
         using_port = None
         for port in range(*port_range):
             if not can_connect_to_port(port):
@@ -194,14 +197,15 @@ class ContextBot:
         # self.app.ui.tabWidget.setCurrentWidget(self.app.ui.tabWidget.findChild(QWidget, 'recording_tab'))  # switch to the recoding widget
         self.qtbot.mouseClick(self.app.recording_tab.StartStopRecordingBtn, QtCore.Qt.MouseButton.LeftButton)  # start the recording
 
-    def start_streams_and_recording(self, num_stream_to_test: int, num_channels: Union[int, Iterable[int]]=1, sampling_rate: Union[int, Iterable[int]]=1, stream_availability_timeout=2 * stream_availability_wait_time * 1e3):
+    def start_streams_and_recording(self, num_stream_to_test: int, num_channels: Union[int, Iterable[int]]=1, sampling_rate: Union[int, Iterable[int]]=1, stream_availability_timeout=2 * 1e3):
         """
         start a given number of streams with given number of channels and sampling rate, and start recording.
         @param num_stream_to_test: int, the number of streams to test
         @param num_channels: int or iterable of int, the number of channels in the stream
         @rtype: object
         """
-
+        from rena.config import stream_availability_wait_time
+        stream_availability_timeout = stream_availability_timeout * stream_availability_wait_time
         if isinstance(num_channels, int):
             num_channels = [num_channels] * num_stream_to_test
         if isinstance(sampling_rate, int):
@@ -242,14 +246,16 @@ def get_random_test_stream_names(num_names: int, alphabet = string.ascii_lowerca
 
 def update_test_cwd():
     print('update_test_cwd: current working directory is', os.getcwd())
-    if os.getcwd().endswith(os.path.join('rena', 'tests')):
-        os.chdir('../')
+    if os.getcwd().endswith('tests'):
+        os.chdir('../rena/')
     elif 'rena' in os.listdir(os.getcwd()):
         os.chdir('rena')
     # else:
     #     raise Exception('update_test_cwd: RenaLabApp test must be run from either <project_root>/rena/tests or <project_root>. Instead cwd is', os.getcwd())
 
 def run_visualization_benchmark(app_main_window, test_context, test_stream_names, num_streams_to_test, num_channels_to_test, sampling_rates_to_test, test_time_second_per_stream, metrics, is_reocrding=False):
+    from rena.utils.buffers import flatten
+
     results = defaultdict(defaultdict(dict).copy)  # use .copy for pickle friendly one-liner
     for n_streams, num_channels, sampling_rate in itertools.product(num_streams_to_test, num_channels_to_test, sampling_rates_to_test):
         stream_names = [test_stream_names.pop(0) for _ in range(n_streams)]
@@ -334,6 +340,8 @@ def visualize_metric_across_test_space_axis(results, axis_index, axis_name, test
 
 
 def run_replay_benchmark(app_main_window, test_context: ContextBot, test_stream_names, num_streams_to_test, num_channels_to_test, sampling_rates_to_test, test_time_second_per_stream, metrics, results_path):
+    from rena.utils.RNStream import RNStream
+
     results = defaultdict(defaultdict(dict).copy)  # use .copy for pickle friendly one-liner
     start_time = time.perf_counter()
     test_axes = {"number of streams": num_streams_to_test, "number of channels": num_channels_to_test, "sampling rate (Hz)": sampling_rates_to_test}
